adding atm pageds and moving to a separate files each page.
